# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Full Express API",
    "version": "1.0.0",
    "description": "A comprehensive Express API with Prisma"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    },
    {
      "url": "https://back-end-production-17b6.up.railway.app/api",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/lessons": {
      "get": {
        "summary": "Get all lessons",
        "tags": [
          "Lessons"
        ],
        "responses": {
          "200": {
            "description": "List of all lessons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lesson"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new lesson",
        "tags": [
          "Lessons"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lesson created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/lessons/{id}": {
      "get": {
        "summary": "Get a lesson by ID",
        "tags": [
          "Lessons"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Lesson ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Lesson found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          },
          "404": {
            "description": "Lesson not found"
          }
        }
      },
      "put": {
        "summary": "Update a lesson",
        "tags": [
          "Lessons"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Lesson ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lesson updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          },
          "404": {
            "description": "Lesson not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a lesson",
        "tags": [
          "Lessons"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Lesson ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Lesson deleted successfully"
          },
          "404": {
            "description": "Lesson not found"
          }
        }
      }
    },
...
    "/resources": {
      "get": {
        "summary": "Get all resources",
        "tags": [
          "Resources"
        ],
        "responses": {
          "200": {
            "description": "List of all resources",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resource"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new resource",
        "tags": [
          "Resources"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resource created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          }
        }
      }
    },
    "/resources/{id}": {
      "get": {
        "summary": "Get a resource by ID",
        "tags": [
          "Resources"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Resource found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "put": {
        "summary": "Update a resource by ID",
        "tags": [
          "Resources"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resource updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a resource by ID",
        "tags": [
          "Resources"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Resource deleted"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Lesson": {
        "type": "object",
        "required": [
          "title",
          "content",
          "order",
          "moduleId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the lesson"
          },
          "title": {
            "type": "string",
            "description": "Lesson title"
          },
          "content": {
            "type": "string",
            "description": "Lesson content"
          },
          "order": {
            "type": "integer",
            "description": "Lesson order within module"
          },
          "moduleId": {
            "type": "string",
            "description": "ID of the module this lesson belongs to"
          },
          "duration": {
            "type": "integer",
            "description": "Lesson duration in minutes"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Lesson creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Lesson last update timestamp"
          }
        }
      },
      "LessonInput": {
        "type": "object",
        "required": [
          "title",
          "content",
          "order",
          "moduleId"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "order": {
            "type": "integer"
          },
          "moduleId": {
            "type": "string"
          },
          "duration": {
            "type": "integer"
          }
        }
      },
...
      "Resource": {
        "type": "object",
        "required": [
          "title",
          "description",
          "type",
          "category",
          "format",
          "thumbnail",
          "author",
          "publishedDate"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the resource"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "downloadUrl": {
            "type": "string"
          },
          "externalUrl": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "publishedDate": {
            "type": "string",
            "format": "date-time"
          },
          "downloads": {
            "type": "integer"
          },
          "rating": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResourceInput": {
        "type": "object",
        "required": [
          "title",
          "description",
          "type",
          "category",
          "format",
          "thumbnail",
          "author",
          "publishedDate"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "downloadUrl": {
            "type": "string"
          },
          "externalUrl": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "publishedDate": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}